cast eob message to (int), for prettyness :P
git-svn-id: http://svn.ratbox.org/svnroot/trunk@5881 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ modules/m_eob.c
@@ -76,8 +76,8 @@ static void ms_eob(struct Client *client_p, struct Client *source_p,
 	DeltaTime = CurrentTime - TheirTime;
 
       sendto_realops_flags(
-       FLAGS_ALL,"*** End of burst from %s (%lu seconds)",
-       source_p->name, DeltaTime );
+       FLAGS_ALL,"*** End of burst from %s (%d seconds)",
+       source_p->name, (int)DeltaTime );
     }
   else
     sendto_realops_flags(FLAGS_ALL,"*** End of burst from %s",
